Map branch

+++ include/global.fieldmap.h
@@ -241,8 +241,8 @@ struct ObjectEvent
              u16 movementDirection:4;
              struct __attribute__((packed))
              {
-                u8 rangeX:4;
-                u8 rangeY:4;
+                u16 rangeX:4;
+                u16 rangeY:4;
              } range;
     /*0x1A*/ u8 fieldEffectSpriteId;
     /*0x1B*/ u8 warpArrowSpriteId;
